+++ EssentialFeediOS/FeedViewController.swift
@@ -1,0 +1,35 @@
+//
+//  FeedViewController.swift
+//  EssentialFeediOS
+//
+//  Created by Jair Moreno Gaspar on 28/12/20.
+//  Copyright © 2020 Jair Moreno G. All rights reserved.
+//
+
+import EssentialFeed
+import UIKit
+
+final public class FeedViewController: UITableViewController {
+    private var loader: FeedLoader?
+    
+    public convenience init(loader: FeedLoader) {
+        self.init()
+        self.loader = loader
+    }
+    
+    public override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        refreshControl = UIRefreshControl()
+        refreshControl?.addTarget(self, action: #selector(load), for: .valueChanged)
+        load()
+    }
+    
+    @objc
+    private func load() {
+        refreshControl?.beginRefreshing()
+        loader?.load { [weak self] _ in
+            self?.refreshControl?.endRefreshing()
+        }
+    }
+}
